More Android compatibility changes
At this point, I've spent a day trying to get Boost to (very slowly)
compile on Windows/Cygwin. It just built on the first attempt on OS X in
less time than it took Windows to unpack the archive file. I'm going
back to OS X for this one, hence the commit.