Add FilterManager, and associated code, to Readium SDK
This change adds the new FilterManager, a class that manages ContentFilters. It can create a ContentFilter when requested. It will create the appropriate ContentFilter based on the data given to it. The FilterManager class itself is an abstract class that can be subclassed into more sophisticated FilterManagers if needed. However, there is already an implementation in this change, called FilterManagerImpl.