Nav: support empty EPUB3 toc. It does this by dropping back to NCX parsing; there are books which contain an EPUB3 nav document containing an empty TOC with only a title, i.e.:
  <nav epub3:type="toc">
    <h2>My Title</h2>
  </nav>

These books have their *real* TOC in the EPUB2 NCX file. Package now handles this by parsing EPUB3, and if the TOC is present but has no children, it will try NCX parsing. If the NCX is not present at this point, it does NOT signal an error-- the assumption is that the EPUB3 nav was correct, and there are simply no chapter markers at all.