Workaround for SFINAE issues on MSVC compiler around `ePub3::async()`.
You'll notice, if you look at the implementation of `std::async()` on Windows 8, that MS doesn't have the non-policy version of the function (or something along those lines-- I don't have a copy here right now). Guess why? ;o)

This commit hides the non-policy version when compiling using MSVC. If you attempt to use it, you should be able to interpret the error code correctly. I went through a whole heck of a lot of template and type fudging to try and find a way to make MSVC actually correctly interpret the parameters, but in the end I couldn't: if a policy is the first argument, MSVC matches the non-policy template; with no policy, it matches the policy template. Le sigh.

Ironic, really, when the whole custom (well, based on Boost) implementation of `<future>` here is specifically to provide support for *Microsoft's* `future::then()` proposal.