First checkin for the implementation of Byte Ranges in Readium SDK (WORK IN PROGRESS)
This is the first checkin for the implementation of Byte Ranges in Readium SDK. In short, the objective of adding Byte Ranges
to Readium SDK is so that there is a way to deal with resources of great size inside an ePUB. For example, you may have an ePUB
file with a 1 GB video file in it, and in that case you want to allow the client app to be able to read just chunks of bytes out of that video,
and not have to put the entire 1 GB video in memory.

This is still a work in progress. There are still code changes that we need to make, plus some more code cleanup that we want to do,
before marking this as concluded. Following checkins will take care of that.