Changing the smart pointer for the vector of byte streams
With the changes we recently made on RDPackage, the code now uses SyncContentStreamForItem() (or SyncByteRangeForItem())
in order to get a byte stream for a given resource. The problem, though, is that those methods return a ByteStreamPtr, which was
defined by Jim to be a shared_ptr<ByteStream>. This way, in order to be able to store them in the m_byteStreamVector member
variable, we need to use a shared_ptr instead of a unique_ptr.