Final chunk of refactoring for Byte Range streams
This checkin includes the final chunk of refactoring to support Byte Range streams. Evidently, there may be further checkins as
we take feedback on these changes, as well as we fix bugs. But this should be the last piece of redesign in terms of the
latest discussions we had around Byte Range streams.

To summarize very quickly, this change removes any change previously made in the Cocoa HTTP Server code. It also removes
the temporary way that the code had to decide between using Byte Ranges or not. Finally, it also refactors how Byte Streams are
built upon for usage by the Cocoa HTTP Server: it first gets the "raw" byte stream, but later switches to the byte streams that
can have ContentFilter objects (and that may support byte ranges).