Changes based on Code Reviews
After code reviews by Daniel and Shane, I'm checking in this changes. However, notice that this checkin is actually
smaller than it looks. One of the things I did in this change is to try to make consistent use of tabs for indentation.
In RDService, tabs have been using consistently to make indentations. With the latest checkins made in RDServices,
some lines of code had spaces instead for indentations. That was making things look funny in GitHub. So, in this
checkin, I tried to switch all the files that I saw with discrepancies to use tabs. Because of that, this checkin is
marked as having a lot of changes, but most of the changes are solely from spaces to tabs.