Fixing an issue for extracting byte ranges when there is no ContentFilter
There was an issue when extracting byte ranges when there is no ContentFilter.
The problem relied in the fact that, when there is no ContentFilter whatsoever,
the byte stream kept by RDPackageResource is a simple ZipFileByteStream.
In that case, the RDPackageResource code should just use the SeekableByteStream
methods in order to extract a given byte range. The current code is pretty similar to
the same fix that Daniel implemented in OS X.