Clean up of latest development efforts
Both Daniel and I have been playing around in the feature/ByteRangeStreams branch,
modifying a bunch of things in order to implement the byte ranges. As part of that
development effort, we commented out some code, added printfs to do debugging,
etc. This checkin cleans up some of that code.