Changes in Android to support font obfuscation
This change tackles the fact that the Launchers need to call
Package::GetFilterChainByteStream() in order to get a byte stream
that has ContentFilter objects hooked up to it. Without that call, you can
get a byte stream, but it does not have any ContentFilter and therefore
things like obfuscated fonts don't work. Notice, however, that this change
does not allow Android to make use of byte ranges yet.