Support for byte ranges in Readium SDK in Android
This checkin includes some changes that will allow Readium SDK to support
byte ranges in Android. In a nutshell, these changes add another native
method to the set of native methods associated with the Package class.
This new method will receive a range, and it will instantiate a
FilterChainByteStreamRange object based upon the request. Then, it
will use that FilterChainByteStreamRange to extract a range of bytes
from a given resource, and it will return those bytes inside a ByteBuffer.
Later, the EpubServer class will put that ByteBuffer inside a
ByteArrayInputStream, and it will pass that to the Nano HTTP Server,
in order to fulfill a request.