Porting fixes for encryption related ContentFilters
When opening an encrypted ePUB file, the Readium SDK will
set up a ContentFilter in order to decrypt that file. However, for some
time now we have been dealing with the issue that a decrypted file is
smaller than an encrypted file. The set of fixes in this checkin (which
come from a separate line of development) fix some issues around
the content length header. As in, in the situations involved in these fixes,
the content length header would be set incorrectly, in the case of an
encrypted file. This would have consequences like, for example,
Quicktime would be left waiting for more bytes to come when there
was none. The fixes here avoid issues like that.