Fixing a small mistake on a forward declaration
With the changes that we had to make to support byte ranges in the Readium SDK,
the original ByteRange struct became bigger and more sophisticated. As such, it
evolved to become a class as opposed to simply a struct. However, the problem was
that there was a number of forward declarations of ByteRange still referring to it as a
struct, which caused a warning since the type was now a class. This change simply
fixes that and removes that warning.