Fixing a number of issues with content length
This checkin takes care of a number of issues regarding content length.
Based on a series of discussions and investigations that we made around
the content length issue, this checkin contains all the changes that we
thought were necessary to make it work properly. Also, most importantly,
as far as I could test it, works will both encrypted and unencrypted content.