Fix for infinite loop while loading an encrypted resource
This checkin fixes the following issue: in some cases, we noticed that
the Readium SDK would try to load a given encrypted resource an
infinite number of times, finally exhausting the memory and crashing
the app. The problem was in how the code was determining that the
cache, in FilterChainByteStream, was filled up already or not. Originally,
it was simply looking if the Zip file entry was at the end. However, the
decryptor could move the internal Zip file pointer around, invalidating
that. So, to avoid that, I added a specific flag to track when the cache
has already been filled up or not.