Fix for uncompressed media playback issue in libzip
The version of the libzip code that has been incorporated into
the Readium SDK has a bug: when it deals with an uncompressed
media file, the code in libzip follows a different code path (since the
file in question has been marked as uncompressed) but the code
will lose track of its correct position, which will result that the
media file will not be played in the Launcher. This fix takes care
of that: it just makes sure that the code keeps track of the right
reading position in the uncompressed file.