Fix for uncompressed media playback issue in libzip
When a file is marked "uncompressed" inside a ZIP file, the
current libzip code in the Readium SDK will take a different code
path. Unfortunately, there is a small bug in that code path that
results that, when dealing with an uncompressed ZIP entry,
the code will lose track of its correct position in the file. This
fix just adds some code so that it can keep track correctly of
the position inside the uncompressed file, and that allows the
playback of the uncompressed media to happen fine. Just a note:
by "uncompressed" we mean uncompressed from the ZIP file
point of view - evidently the media file, if it is a MPEG-4 file for
example, will be highly compressed.