Creating a new architecture for error reporting in ContentFilters
ContentFilters are one of the mechanisms that the Readium SDK
uses for extensibility. However, up to this point, there was not a
clear way to report an error inside a ContentFilter. For example,
if there is any error inside the FilterData() function, there would be
no clear way to report that back to the app. This change creates
a new architecture through which the code inside a given
ContentFilter will be able to report an error back to the app, and
the app will be able to do something about that (like displaying
the error to the user).