Adding code to the JNI layer to deal with ContentFilter errors
Recently, I implemented a change in the Readium SDK that allows
ContentFilters to report errors. That is important because the
ContentFilters are the basic mechanism for extending the Readium
SDK, and it is expected that people all over the world will start
writing their own ContentFilters. Given that, it is necessary that
there is a mechanism for reporting when there is an error in a given
ContentFilter. The basic mechanism in the Readium SDK has already
been implemented, but this change adds the extra glue code in the
JNI layer so that Android apps can receive those error messages.